% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abridged}
\alias{abridged}
\title{Abridged Rubisco kinetics data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 243 rows and 15 columns.
}
\usage{
abridged
}
\description{
Abridged Rubisco kinetics data
}
\keyword{datasets}
