% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHScale.R
\name{check_DHScale}
\alias{check_DHScale}
\title{Validator for the S3 class 'DHScale'}
\usage{
check_DHScale(the_DHScale)
}
\arguments{
\item{the_DHScale}{the 'DHScale' instance to be checked.}
}
\value{
the supplied DHScale instance, if all checks pass.
}
\description{
Throw error if any scaling constants are non-numeric, or if name
is not a character string.
}
\examples{
bad_DHScale <- new_DHScale(NA, 41, 26, -13.8, name="bad") # bad instance
try(check_DHScale(bad_DHScale)) # produces error

}
