% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme.R
\name{check_enzyme}
\alias{check_enzyme}
\title{Validator for the S3 class 'enzyme'}
\usage{
check_enzyme(the_enzyme)
}
\arguments{
\item{the_enzyme}{the 'enzyme' instance to be checked.}
}
\value{
the supplied enzyme instance, if all checks pass.
}
\description{
Throw error if any kinetic values or temperature are non-numeric, or if name
is not a character string.
}
\examples{
bad_enzyme <- new_enzyme(NA, 144, 394, 10.8, name="bad") # bad instance
try(check_enzyme(bad_enzyme)) # produces error
}
