% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cite_data}
\alias{cite_data}
\title{Cite data collected in the package.}
\usage{
cite_data(identifier, data = "abridged")
}
\arguments{
\item{identifier}{The identifier of the data.}

\item{data}{The data table from which the identifier was pulled.}
}
\value{
For data from a single study, this returns the citation for that
study. For data averaged or compiled by this study, returns a note to
cite this package.
}
\description{
Cite data collected in the package.
}
\examples{
print(EXAMPLE PENDING)
}
