% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cite_Rbc}
\alias{cite_Rbc}
\title{Cite data collected in the package.}
\usage{
cite_Rbc(identifier, type = "full")
}
\arguments{
\item{identifier}{The identifier of the data. Takes a single identifier or a
list of identifiers.}

\item{type}{The type of citation desired. Can take "full", which will return
the full citation, or "short", which returns the pmid or doi, if available.}
}
\value{
For data from a single study, this returns the citation for that
study. For data averaged or compiled by this study, returns a note to
cite this package.
}
\description{
Cite data collected in the package.
}
\examples{
cite_Rbc("average_1B_all")
cite_Rbc(c("japonica_Sage_2002b","japonica_orr_2016_dH","average_1B_C3_warm_dH"))
cite_Rbc("breve_banda_2020",type="short")
}
