% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{CO2_dependence}
\alias{CO2_dependence}
\title{Function factory to model an enzyme's carbon output}
\usage{
CO2_dependence(enzyme, DHScale, pathway)
}
\arguments{
\item{enzyme}{An 'enzyme' object. (See enzyme class for more info.)}

\item{DHScale}{A 'DHScale' object. (See DHScale class for more info.)}

\item{pathway}{One of 'gross', 'canon', or 'alt'. This option allows you to
pick a stoichiometry for the phosphoglycolate salvage (PGS) pathway. If you
do not want to incorporate PGS, use the 'gross' option, while 'canon'
applies the C2 cycle/glycerate pathway stoichiometry (1 CO2 for
every 2 2-phosphoglycolate created). 'Alt' corresponds to the oxalyl-CoA or
malate cycle pathways (2 CO2 for every 2-phosphoglycolate created).}
}
\value{
A new function that models carbon output with three variables: CO2
(in uM), O2 (in uM), and temperature (in celsius).
}
\description{
Function factory to model an enzyme's carbon output versus temperature
and CO2/O2 concentration. Uses a 'DHScale' object to scale each kinetic
parameter from an 'enzyme' object in regards to temperature. Based on the
stoichiometry of known phosphoglycolate salvage pathways, it also takes
into account the carbon lost due to the Rubisco oxygenation reaction based
on user input.
}
\examples{
pro_enzyme <- new_enzyme(6.6, 144, 394, 10.8, name="pro")
bacIA_DHScale <- new_DHScale(47.2,40.8,26.7,-21.8,name="bacIA")
pro_model <- CO2_dependence(pro_enzyme, bacIA_DHScale, "canon") # sets up model
pro_model(126,54,15) # prints carbon fixation at 15 deg C, 126 uM CO2, 54 uM O2
}
