% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHScale.R
\name{DHScale}
\alias{DHScale}
\title{Create an DHScale instance by looking up the relevant data in database}
\usage{
DHScale(id_name, id_col = "identifier", scale_name = NULL, data = NULL)
}
\arguments{
\item{id_name}{the name of identifier used to pull out a single record
out of the database}

\item{id_col}{the name of the column from which the id_name will be searched
for within the database}

\item{scale_name}{the name of the DHScale instance. If NULL the value is
defaulted to id_name}

\item{data}{the database to search from. Can be "averaged", NULL, or a
data.frame. If NULL or "averaged", the averaged table is searched.
NOTE: if a custom data.frame is supplied, it is assumed to have columns
named "kcat_dH", "Kc_dH", "Ko_dH", and "S_dH".}
}
\value{
a new DHScale instance
}
\description{
Create an DHScale instance by looking up the relevant data in database
}
\examples{
cyano <- DHScale("cyano_1Ac") # retrieved from abridged database
}
