% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme.R
\name{Enzyme}
\alias{Enzyme}
\title{Create an Enzyme instance by looking up the relevant data in database}
\usage{
Enzyme(id_name, id_col = "identifier", enzyme_name = NULL, data = NULL)
}
\arguments{
\item{id_name}{the name of identifier used to pull out a single record
out of the database}

\item{id_col}{the name of the column from which the id_name will be searched
for within the database}

\item{enzyme_name}{the name of the enzyme instance. If NULL the value is
defaulted to id_name}

\item{data}{the database to search from. Can be "abridged", "comprehensive",
NULL, or a data.frame. If NULL, the abridged table is first searched and
if no entry is found the comprehensive table is searched.
NOTE: if a custom data.frame is supplied, it is assumed to have columns
named "kcat_val", "Kc_val", "Ko_val", and "S_val". In addition, it may
have optional columns "temp", "kcat_T", "Kc_T", "Ko_T", "S_T", and "PGS"}
}
\value{
a new Enzyme instance
}
\description{
Create an Enzyme instance by looking up the relevant data in database
}
\examples{
biuncialis <- Enzyme("biuncialis_prins_2016") # retrieved from abridged database
# specify enzyme using species name
vavilovii <- Enzyme("vavilovii", "species", data="comprehensive")
}
