% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\encoding{UTF-8}
\name{kinetics_T_dep}
\alias{kinetics_T_dep}
\title{Function factory for calculating temperature-dependent kinetics}
\usage{
kinetics_T_dep(std, delta_H, temp = 25)
}
\arguments{
\item{std}{Measured value}

\item{delta_H}{ΔH value for this kinetic parameter for this organism type
(in kJ/mol).}

\item{temp}{Temperature at which param 'std' was measured. Defaults to 25°C.}
}
\value{
A new function that scales the supplied kinetic parameter with
temperature (in celsius).
}
\description{
Calculates the non-supplied constant (c) for an Arrhenius-type temperature
scaling function (as seen in Galmés et al. 2016). Uses measured values and
the temperature they were measured at, along with a supplied temperature
scaling constant (also called the activation energy, ΔH (in kJ/mol)).
}
\examples{
kcat_proteo_Ia <- kinetics_T_dep(15, 44.7) # create new function
kcat_proteo_Ia(30) # evaluate the kinetic parameter at a different temperature

}
