% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_non_NA}
\alias{last_non_NA}
\title{Return the last non-NA value from a vector.}
\usage{
last_non_NA(x)
}
\arguments{
\item{x}{the vector to extract values from.}
}
\value{
The last non-NA value, or NA if none exists.
}
\description{
Return the last non-NA value from a vector.
}
