% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{make_4D_grid}
\alias{make_4D_grid}
\title{Construct a 4D grid}
\usage{
make_4D_grid(func, x_seq, y_seq, z_seq, var_names = c("x", "y", "z", "f"))
}
\arguments{
\item{func}{The function to apply on the grid points to obtain the
value of the dependent variable.}

\item{x_seq}{A vector of coordinates for the first independent variable.}

\item{y_seq}{A vector of coordinates for the second independent variable.}

\item{z_seq}{A vector of coordinates for the third independent variable.}

\item{var_names}{Names for the 3 independent and 1 dependent variables,
in a single character vector.}
}
\value{
A 4D grid as a 4-element list of 3-dimensional arrays
}
\description{
Construct a 4D grid, where the first 3 coordinates are independent
variables, while the 4th coordinate is dependent variable arising from
evaluating a function against the values of the first 3 coordinates.
}
\details{
The grid is represented by a 4-element list, where each member is a
3-dimensional array representing grid points. The first array is the
value of first independent variable on each grid point, etc., and the
fourth array is the value of function evaluated at each grid point.
}
\examples{
f <- function(x, y, z) { x + y * z }
g1 <- make_4D_grid(f, 1:3, c(-2,2), seq(4, 10, 2))
g1
}
