% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{median_if_num}
\alias{median_if_num}
\title{Compute the median of a vector if the vector is numerical or logical.
Otherwise, return the common value if all entries take the same value,
or NA in any other cases.}
\usage{
median_if_num(x, na.rm = TRUE)
}
\arguments{
\item{x}{the vector to compute average from.}

\item{na.rm}{whether NA's are ignored for the purpose of computing average
or for checking consistency between entries.}
}
\value{
The average or common value, if exists.
}
\description{
Compute the median of a vector if the vector is numerical or logical.
Otherwise, return the common value if all entries take the same value,
or NA in any other cases.
}
