% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_entries}
\alias{merge_entries}
\title{Merge (combine) several entries from a data.frame.}
\usage{
merge_entries(
  data,
  id_col,
  id_vals,
  method,
  new_id = NA,
  const_cols = NULL,
  keep_merged = TRUE,
  keep_unmerged = TRUE
)
}
\arguments{
\item{data}{The data.frame from which entries are pulled.}

\item{id_col}{The column from which the relevant entries are identified.}

\item{id_vals}{The values from id_col for which the corresponding entries
will be pulled.}

\item{method}{The method to combine entries. Can be "average" or "last".}

\item{new_id}{The new identifier (for id_col) for the merged entry.}

\item{const_cols}{The columns to check for consistency. Error is thrown
if the values from the pulled entries are not identical for any such
columns.}

\item{keep_merged}{Whether to keep the original entries that are pulled
and merged.}

\item{keep_unmerged}{Whether to keep the original entries that have NOT
been touched.}
}
\description{
Merge (combine) several entries from a data.frame.
}
\examples{
df <- data.frame(
  id = 1:4,
  x = c(1, 2, NA, 3),
  y = c("this", "this", NA, "that")
)

df2 <- merge_entries(df, "id", c(1,4), "average", new_id = "new")
df2

df3 <- merge_entries(df, "id", c(4, 3, 1), "last", keep_merged = FALSE)
df3

try(merge_entries(df, "id", c(4, 3, 1), "last", const_cols="y")) # expects error
}
