% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHScale.R
\name{modify_DHScale}
\alias{modify_DHScale}
\title{Create a new instance of S3 class 'DHScale', starting from a previous
instance}
\usage{
modify_DHScale(
  the_DHScale,
  kcat_dH = NULL,
  kc_dH = NULL,
  ko_dH = NULL,
  s_dH = NULL,
  name = NULL
)
}
\arguments{
\item{the_DHScale}{the previous 'DHScale' instance to base the new
instance on.}

\item{kcat_dH}{if not NULL, the kcat_dH value for the new instance}

\item{kc_dH}{if not NULL, the kc_dH value for the new instance}

\item{ko_dH}{if not NULL, the ko_dH value for the new instance}

\item{s_dH}{if not NULL, the s_dH value for the new instance}

\item{name}{if not NULL, the identifier ("name") for the new instance}
}
\value{
a new instance of the S3 class 'DHScale'.
}
\description{
If any argument (other than \code{the_DHScale}, which is required) is \code{NULL},
the corresponding instance attribute will copy its value from the previous
instance. Otherwise, the input value will become the value of the
corresponding attribute in the new instance. For detailed explanation of
each 'DHScale' attribute, see the documentation of \code{new_DHScale()}
}
\examples{
bacIA_DHScale <- new_DHScale(47.2,40.8,26.7,-21.8,name="bacIA") # new instance
print(bacIA_DHScale) # printout DHScale info
tmrIA_DHScale <- modify_DHScale(bacIA_DHScale,kcat_dH=32,name="tmr_IA") # modify instance
print(tmrIA_DHScale) # printout modified DHScale info

}
