% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHScale.R
\name{new_DHScale}
\alias{new_DHScale}
\title{Constructor for the S3 class 'DHScale'.}
\usage{
new_DHScale(kcat_dH, Kc_dH, Ko_dH, S_dH, name = "")
}
\arguments{
\item{kcat_dH}{The temperature scaling constant of kcat for the carboxylation
reaction, also known as the activation energy (in kJ/mol).}

\item{Kc_dH}{The temperature scaling constant of Kc (in kJ/mol).}

\item{Ko_dH}{The temperature scaling constant of Ko (in kJ/mol).}

\item{S_dH}{The temperature scaling constant of Sc/o (in kJ/mol).}

\item{name}{An identifier ("name") of the scale instance.}
}
\value{
A new 'DHScale' instance.
}
\description{
This class collects temperature scaling constants for each of the four
kinetic parameters. These scaling constants are also called the activation
energy of that kinetic parameter, and are typically denoted with ΔH. In
combination with an enzyme-class object, a DHScale-class object can be used
to create a function that models an enzyme's carbon output with regards to
temperature and CO2/O2 concentration.
}
\examples{
new_DHScale(40.1, 38.8, 26.7, -27.4, name = "Form Ia, cyanobacteria")
}
