% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme.R
\name{new_enzyme}
\alias{new_enzyme}
\title{Constructor for the S3 class 'enzyme'.}
\usage{
new_enzyme(
  kcat_val,
  Kc_val,
  Ko_val,
  S_val,
  temp = 25,
  kcat_T = NULL,
  Kc_T = NULL,
  Ko_T = NULL,
  S_T = NULL,
  T_max = 55,
  PGS = NA,
  name = ""
)
}
\arguments{
\item{kcat_val}{The measured value of kcat for the Rubisco carboxylase
reaction (in s^-1). kcat represents the number of reactions per second a
single active site can perform. The temperature at which it was measured
can be found in kcat_T (if present) or temp.}

\item{Kc_val}{The measured value of Kc (in uM). Kc is the Michaelis constant
of Rubisco for CO2, or its affinity for CO2. The temperature at which it
was measured can be found in Kc_T (if present) or temp.}

\item{Ko_val}{The measured value of Ko (in uM). Ko is the Michaelis constant
of Rubisco for O2, or its affinity for O2. The temperature at which it
was measured can be found in Ko_T (if present) or temp.}

\item{S_val}{The measured value of Sc/o (unitless). Sc/o is the specificity of
Rubisco for CO2 versus O2. The temperature at which it was measured can be
found in S_T (if present) or temp.}

\item{temp}{The common temperature (in celsius) at which the kinetic
constants are measured. Can be overridden by the more specialized
arguments kcat_T, Kc_T, Ko_T, and S_T. If NULL these 4 arguments are
used to specify temperatures.}

\item{kcat_T}{The temperature (in celsius) at which kcat_val is measured.}

\item{Kc_T}{The temperature (in celsius) at which Kc_val is measured.}

\item{Ko_T}{The temperature (in celsius) at which Ko_val is measured.}

\item{S_T}{The temperature (in celsius) at which S_val is measured.}

\item{T_max}{The upper thermal limit (in celsius) of the enzyme.}

\item{PGS}{The default phosphoglycolate salvage (PGS) pathway of the enzyme.
Can be NA.}

\item{name}{An identifier ("name") of the enzyme instance.}
}
\value{
A new enzyme instance.
}
\description{
The class 'enzyme' is designed to encapsulate the information necessary to
describe a particular measured enzyme's response to CO2, O2, and temperature.
All of Rubisco's key kinetic parameters are used as inputs, along with the
temperature at which they were measured. In combination with a DHScale-class
object, an enzyme-class object can be used to create a function that models
a Rubisco enzyme's carbon output with regards to temperature and CO2/O2
concentration.
}
\examples{
new_enzyme(6.6, 144, 394, 10.8, name="pro")
}
