% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{permute_4D_grid}
\alias{permute_4D_grid}
\title{Permute the independent variables in a 4D grid created by make_4D_grid().}
\usage{
permute_4D_grid(grid, perm)
}
\arguments{
\item{grid}{The original grid to be permuted.}

\item{perm}{the subscript permutation vector, a permutation of the integers
c(1, 2, 3).}
}
\value{
A permuted 4D grid
}
\description{
In the new grid, every element in the list is a permuted array. Moreover,
the order of the independent variables in the list is also permuted.
}
\examples{
f <- function(x, y, z) { x + y * z }
g1 <- make_4D_grid(f, 1:3, c(-2,2), seq(4, 10, 2))
g2 <- permute_4D_grid(g1, c(2,3,1))
g2
}
