% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_slice_3D}
\alias{plot_slice_3D}
\title{Create a plot of the dependent variable in a 3D slice}
\usage{
plot_slice_3D(
  grid_slice,
  contours = NULL,
  cmin = NULL,
  cmax = NULL,
  dims = NULL,
  colors = "default",
  NA_color = "grey",
  contour_col = "black",
  xlabel = "",
  ylabel = "",
  clabel = c("Carbon", "(C/s)"),
  lwd = 2,
  ...
)
}
\arguments{
\item{grid_slice}{The 3D slice to be plotted}

\item{contours}{The values of dependent variable at which the contours
are drawn, as a single numeric vector. If NULL no contours are produced.}

\item{cmin}{The lower bound for the color mapping. If NULL defaults to the
minimum value of the contour/color variable.}

\item{cmax}{The upper bound for the color mapping. If NULL defaults to the
maximum value of the contour/color variable.}

\item{dims}{Length 3 integer vector, e.g., c(1,2,4), that specify the
horizontal, vertical, and contour/color variables of the plot. If NULL,
the function will attempt to automatically determine horizontal coordinate
and the vertical coordinate, while the last element of the slice is
treated as the dependent variable. Note that the horizontal coordinate
should ALWAYS correspond the first index of the two-dimension arrays,
and the vertical coordinates should ALWAYS be the second index. (Use
transpose_3D_slice() to modify which variables are horizontal/vertical)}

\item{colors}{The vector of colors used for rendering the false color
tiles of the plot. If the string "default" is supplied instead, the
default blue-white-red color scale is used, where the color is bluer
the more negative the dependent variable, redder the more positive the
dependent variable, and which white is anchored at 0.}

\item{NA_color}{The color used to represent undefined or out-of-range values}

\item{contour_col}{The color for the contour lines}

\item{xlabel}{The label for the horizontal axis}

\item{ylabel}{The label for the vertical axis}

\item{clabel}{The label for the color bar. Defaults to c("Carbon", "(C/s)")}

\item{lwd}{The linewidth of contour lines}

\item{...}{Addtional arguments are passed to plot3D::image2D(), which is the
"bottom" plot created by this function (to be overlaid by the contour plot)}
}
\value{
No explicit return (plot generated as side effect)
}
\description{
Create a plot of the dependent variable (4-th index) in a 3D slice
(created using, e.g., slice_4D_grid()) against its two non-constant
independent dimensions. In the plot, the dependent variable is
represented by contours and false color, while the two independent
variables form the horizontal and the vertical axes.
}
\examples{
# create 3D slices
f <- function(x, y, z) { x + y * z }
g1 <- make_4D_grid(f, seq(1, 3, 0.1), seq(-2, 2, 0.2), seq(4, 10, 0.2))
s1 <- slice_4D_grid(g1, 2, 2)
# plot the 3D slice
plot_slice_3D(s1, contours=seq(5, 25, 2.5), dims=c(1, 3, 4))
}
