% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHScale.R
\name{print.DHScale}
\alias{print.DHScale}
\title{Customized print for S3 class 'DHScale'}
\usage{
\method{print}{DHScale}(x, unicode = TRUE, ...)
}
\arguments{
\item{x}{the 'DHScale' instance to be printed.}

\item{unicode}{whether to print using unicode characters.}

\item{...}{the remaining arguments are ignored.}
}
\value{
No explicit return. Information of the 'DHScale' instance is printed
to stdout.
}
\description{
Customized print for S3 class 'DHScale'
}
\examples{
ia_cyano_DHScale <- new_DHScale(40.1, 38.8, 26.7, -27.4, name = "Form Ia, cyanobacteria")
print(ia_cyano_DHScale) # printout DHScale info

}
