% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme.R
\name{print.enzyme}
\alias{print.enzyme}
\title{Customized print for S3 class 'enzyme'}
\usage{
\method{print}{enzyme}(x, unicode = TRUE, ...)
}
\arguments{
\item{x}{the 'enzyme' instance to be printed.}

\item{unicode}{whether to print using unicode characters.}

\item{...}{the remaining arguments are ignored.}
}
\value{
No explicit return. Information of the enzyme instance is printed
to stdout.
}
\description{
Customized print for S3 class 'enzyme'
}
\examples{
pro_enzyme <- new_enzyme(6.6, 144, 394, 10.8, name="pro") # new instance
print(pro_enzyme) # print out enzyme info
}
