% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refs_to_citation}
\alias{refs_to_citation}
\title{References table}
\format{
\subsection{'refs_to_citation'}{

A data frame with 150 rows and 3 columns:
\describe{
\item{short_ref}{Short reference}
\item{pmid_or_doi}{The pmid number or the doi}
\item{citation}{Full citation}
}
}
}
\source{
compiled by the authors
}
\usage{
refs_to_citation
}
\description{
A table containing a list of the citation information and pmid/doi numbers to help with citing data from the package.
}
\keyword{datasets}
