% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rubisco_kinetics}
\alias{Rubisco_kinetics}
\title{Comprehensive Rubisco kinetics data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 960 rows and 30 columns.
}
\usage{
Rubisco_kinetics
}
\description{
Comprehensive Rubisco kinetics data
}
\keyword{datasets}
