% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme.R
\name{search_alias}
\alias{search_alias}
\title{Search for data of Rubisco enzyme alias from the database}
\usage{
search_alias(string, data = NULL, match = "complete")
}
\arguments{
\item{string}{a string to search for in the database}

\item{data}{the source data to search from. Can be "abridged", "comprehensive",
or NULL. If NULL, the search will proceed from abridged to comprehensive
until a match is found}

\item{match}{whether to require complete match. Can be "complete" or "partial"}
}
\value{
a dataframe containing matching entries.
}
\description{
Data is pulled either from the abridged table or the comprehensive table,
both of which are included in this package
}
