% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHScale.R
\name{search_DHScale}
\alias{search_DHScale}
\title{Search for temperature dependence DHScale from the database}
\usage{
search_DHScale(string, level = NULL, data = NULL, match = "complete")
}
\arguments{
\item{string}{a string to search for in the database}

\item{level}{the level at which to search. Can be "species", "genus",
"taxonomy", "form", or NULL. If NULL, the search will proceed from genus
to species to taxonomy to form until a match is found}

\item{data}{the source data to search from. Can be "averaged", "abridged",
or NULL. If NULL, the search will proceed from abridged to averaged
until a match is found}

\item{match}{whether to require complete match. Can be "complete" or "partial"}
}
\value{
a dataframe containing matching entries.
}
\description{
Data is pulled either from the averaged table or the abridged table,
both of which are included in this package
}
\examples{
Aegilops <- search_DHScale("Aegilops", level="genus") # complete genus search
print(Aegilops) # printout search results
}
