% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme.R
\name{search_enzyme}
\alias{search_enzyme}
\title{Search for data of Rubisco enzyme from the database}
\usage{
search_enzyme(string, level = NULL, data = NULL, match = "complete")
}
\arguments{
\item{string}{a string to search for in the database (case insensitive)}

\item{level}{the taxonomic level to search. Can be "alias", "genus", "species",
"form", "taxonomy", or NULL. If NULL, the search will proceed from alias to
genus to species to taxonomy to form until a match is found}

\item{data}{the source data to search from. Can be "abridged", "comprehensive",
or NULL. If NULL, the search will proceed from abridged to comprehensive
until a match is found}

\item{match}{whether to require complete match. Can be "complete" or "partial"}
}
\value{
a dataframe containing matching entries.
}
\description{
Data is pulled either from the abridged table or the comprehensive table,
both of which are included in this package
}
\examples{
Aegilops <- search_enzyme("Aegilops", level="genus") # complete genus search
print(Aegilops) # printout search results
}
