% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{slice_4D_grid}
\alias{slice_4D_grid}
\title{Create a 3D slice of 4D grid created by make_4D_grid().}
\usage{
slice_4D_grid(grid, dim, val, approx = TRUE, tol = 0.001)
}
\arguments{
\item{grid}{The original grid to be sliced.}

\item{dim}{The dimension to be sliced. Its corresponding coordinates will
take constant values in the resulting slice.}

\item{val}{The constant value that the dim-th coordinates will take in the
resulting slice.}

\item{approx}{Whether approximate value matching is allowed.}

\item{tol}{the tolerance for approximate value matching.}
}
\value{
A 3D slice of the supplied 3D grid.
}
\description{
The resulting slice remains a list of 4 elements, but each element except
the dimension being sliced will be a 2-dimensional array. For the dimension
being sliced, the element will be a scalar showing the constant value of the
sliced coordinate.
}
\examples{
f <- function(x, y, z) { x + y * z }
g1 <- make_4D_grid(f, 1:3, c(-2,2), seq(4, 10, 2))
s1 <- slice_4D_grid(g1, 2, 2)
s1
}
