% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp_dep_abridged}
\alias{temp_dep_abridged}
\title{Abridged Delta H temperature scaling data}
\format{
\subsection{'temp_dep_abridged'}{

A data frame with 182 rows and 13 columns:
\describe{
\item{identifier}{Unique entry ID}
\item{genus}{Genus}
\item{species}{Species}
\item{subspecies}{Subspecies, variant, or cultivar}
\item{kcat_dH}{kcat temperature dependence parameter (kJ/mol)}
\item{Kc_dH}{Kc temperature dependence parameter (kJ/mol)}
\item{Ko_dH}{Ko temperature dependence parameter (kJ/mol)}
\item{S_dH}{S temperature dependence parameter (kJ/mol)}
\item{form}{Form of Rubisco}
\item{taxonomy}{Taxonomy}
\item{group}{Broader phylogenetic group}
\item{note}{Note}
\item{short_ref}{Short reference}
}
}
}
\source{
compiled by the authors from a literature search
}
\usage{
temp_dep_abridged
}
\description{
A dataset of Rubisco temperature dependence data compiled from a literature search.
}
\keyword{datasets}
