% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp_dep_averaged}
\alias{temp_dep_averaged}
\title{Averaged Delta H temperature scaling data}
\format{
\subsection{'temp_dep_averaged'}{

A data frame with 9 rows and 13 columns:
\describe{
\item{identifier}{Unique entry ID}
\item{genus}{Genus}
\item{species}{Species}
\item{subspecies}{Subspecies, variant, or cultivar}
\item{kcat_dH}{kcat temperature dependence parameter (kJ/mol)}
\item{Kc_dH}{Kc temperature dependence parameter (kJ/mol)}
\item{Ko_dH}{Ko temperature dependence parameter (kJ/mol)}
\item{S_dH}{S temperature dependence parameter (kJ/mol)}
\item{form}{Form of Rubisco}
\item{taxonomy}{Taxonomy}
\item{group}{Broader phylogenetic group}
\item{note}{Note}
\item{short_ref}{Short reference}
}
}
}
\source{
compiled by the authors, mostly from Galm\enc{é}{e}s et al. 2016
}
\usage{
temp_dep_averaged
}
\description{
A dataset of average Rubisco temperature scaling data per taxonomic group and form, as described in Section 5.3 of the "Designing Custom Enzymes" vignette.
}
\keyword{datasets}
