% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{transpose_3D_slice}
\alias{transpose_3D_slice}
\title{Transpose the non-constant independent variables in a 3D grid created by
slice_4D_grid().}
\usage{
transpose_3D_slice(slice)
}
\arguments{
\item{slice}{The original slice to be transposed.}
}
\value{
A transposed 3D slice.
}
\description{
In the new slice, every non-constant element in the list is transposed.
Moreover, the order of the 2 non-constant independent variables in the list
is also swapped.
}
\examples{
f <- function(x, y, z) { x + y * z }
g1 <- make_4D_grid(f, 1:3, c(-2,2), seq(4, 10, 2))
s1 <- slice_4D_grid(g1, 2, 2)
s2 <- transpose_3D_slice(s1)
s2
}
