% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme.R
\name{modify_enzyme}
\alias{modify_enzyme}
\title{Create a new instance of S3 class 'enzyme' starting from a previous instance}
\usage{
modify_enzyme(
  the_enzyme,
  kcat_val = NULL,
  kc_val = NULL,
  ko_val = NULL,
  s_val = NULL,
  kcat_T = NULL,
  kc_T = NULL,
  ko_T = NULL,
  s_T = NULL,
  name = NULL
)
}
\arguments{
\item{the_enzyme}{the previous 'enzyme' instance to base the new instance on.}

\item{kcat_val}{if not NULL, the kcat_val value for the new instance}

\item{kc_val}{if not NULL, the kc_val value for the new instance}

\item{ko_val}{if not NULL, the ko_val value for the new instance}

\item{s_val}{if not NULL, the s_val value for the new instance}

\item{kcat_T}{if not NULL, the kcat_T value for the new instance}

\item{kc_T}{if not NULL, the kc_T value for the new instance}

\item{ko_T}{if not NULL, the ko_T value for the new instance}

\item{s_T}{if not NULL, the s_T value for the new instance}

\item{name}{if not NULL, the identifier ("name") for the new instance}
}
\value{
a new instance of the S3 class 'enzyme'.
}
\description{
If any argument (other than \code{the_enzyme}, which is required) is \code{NULL},
the corresponding instance attribute will copy its value from the previous
instance. Otherwise, the input value will become the value of the
corresponding attribute in the new instance. For detailed explanation of
each 'enzyme' attribute, see the documentation of \code{\link[=new_enzyme]{new_enzyme()}}
}
\examples{
pro_enzyme <- new_enzyme(6.6, 144, 394, 10.8, name="pro") # new instance
print(pro_enzyme) # printout enzyme info
pro_worse_enzyme <- modify_enzyme(pro_enzyme, kcat_val=2, kc_val=310, name="pro_worse")
print(pro_worse_enzyme) # printout modified enzyme info
}
