% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rubisco_25C}
\alias{Rubisco_25C}
\title{Comprehensive Rubisco kinetics data}
\format{
An object of class \code{data.frame} with 869 rows and 31 columns.
}
\usage{
Rubisco_25C
}
\description{
Comprehensive Rubisco kinetics data
}
\keyword{datasets}
