% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rubisco_abridged}
\alias{Rubisco_abridged}
\title{Abridged Rubisco kinetics data}
\format{
An object of class \code{data.frame} with 258 rows and 17 columns.
}
\usage{
Rubisco_abridged
}
\description{
Abridged Rubisco kinetics data
}
\keyword{datasets}
