% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rubisco_aliases}
\alias{Rubisco_aliases}
\title{Rubisco alias table}
\format{
An object of class \code{data.frame} with 72 rows and 4 columns.
}
\usage{
Rubisco_aliases
}
\description{
Rubisco alias table
}
\keyword{datasets}
