% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp_dep_abridged}
\alias{temp_dep_abridged}
\title{Abridged Delta H temperature scaling data}
\format{
An object of class \code{data.frame} with 111 rows and 12 columns.
}
\usage{
temp_dep_abridged
}
\description{
Abridged Delta H temperature scaling data
}
\keyword{datasets}
