% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp_dep_averaged}
\alias{temp_dep_averaged}
\title{Averaged Delta H temperature scaling data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 10 columns.
}
\usage{
temp_dep_averaged
}
\description{
Averaged Delta H temperature scaling data
}
\keyword{datasets}
